+++ examples/string_transformer_workflow.py
@@ -19,6 +19,20 @@ class StringTransformer(pyflow.Workflow):
         for s in workflow_input:
             upcased.append(self.string_upcase(s))
 
+        # is_replaying is True if this or some other decider process already executed this line
+        if not self.swf.is_replaying:
+            print "First time hitting this line"
+
+        # Another way to run something only once, also propagates the result.
+        def run_once(x):
+            print "run_once({!r}) called".format(x)
+            return x + 1
+
+        run_once_fut = self.swf.invoke_once(run_once, 42)
+        result = run_once_fut.result()
+        if not self.swf.is_replaying:
+            print "run_once result = {!r}".format(result)
+
         # demonstrate error handling
         try:
             # pass a number where a string is expected
@@ -62,7 +76,7 @@ class StringConcatter(pyflow.Workflow):
 
 def main():
     logging.basicConfig()
-    pyflow.logger.setLevel(logging.DEBUG)
+    pyflow.logger.setLevel(logging.INFO)
 
     workflows = [StringTransformer, StringConcatter]
     domain = 'SWFSampleDomain'